Make `genproto` a `cfg` flag instead of a feature
Previously, the build dependencies of the `genproto` feature weren't
optional, so they were built even when not required, leading to some
breakage downstream.

However, `genproto` arguably generally shouldn't be a user-facing feature of this crate,
but is rather better hidden behind a `cfg` flag. Here, we do exactly
that, also fixing the issue mentioned above.