Allow to set `aad` in `StorableBuilder::{build,desconstruct}`
Previously, we omitted setting `ChaCha20Poly1305RFC`'s `aad` field,
which had the `tag` not commit to any particular key. This would allow a
malicious VSS provider to substitute blobs stored under a different key
without the client noticing.

Here, we allow to set the `aad` (which usually should be the full key
under which we expect the `Storable` to get stored). This change is
backwards compatible if users provide a static `aad: &[]` argument.
Otherwise, users will need to find an upgrade path client-side.