Utilize `reqwest::retry::Builder` to implement HTTP retries
We here make use of `reqwest`'s `retry` functionlity that was introduced
with its recent v0.12.23 release. As we can't fully replace the old
'application-level' behavior just yet, we configure it to only apply for
error that arent INTERNAL_SERVER_ERROR, which is the status VSS uses to
send error responses.

We set some default values here, but note that these can always can be
overridden by the user when using the `from_client` constructor.